extracted MessageType to own file
